+++ test/setup/connect_setup.py
@@ -7,7 +7,7 @@ import logging
 import os
 
 # use the perftest fuzzbucket instance since it already has all the deps
-alias = "ubuntu22-publishing-client-2024.05"
+alias = "ubuntu22-publishing-client-2024.09"
 box_name = "connect-publishing-client"
 list_command = "fuzzbucket-client -j list"
 create_command = "fuzzbucket-client create -c -S 20 -t m5.2xlarge " + alias + " -n " + box_name
@@ -68,12 +68,8 @@ def connect_ready(box_name, max_attempts, interval):
                     update_config="fuzzbucket-client ssh " + box_name + " " + ssh_options + " sudo sed -i 's/CONNECT_IP/" + connect_box + "/g' /etc/rstudio-connect/rstudio-connect.gcfg"
                     restart_connect = "fuzzbucket-client ssh " + box_name + " " + ssh_options + " sudo systemctl restart rstudio-connect"
                     logging.info("Installing Connect on " + connect_box)
-                    # kill any apt locks before installing Connect
-                    # if no processes are found it will continue but will throw an error
-                    try:
-                        subprocess.check_output("fuzzbucket-client ssh -q " + box_name + " " + ssh_options + " sudo lsof -t /var/lib/dpkg/lock-frontend | xargs kill", shell=True, text=True)
-                    except subprocess.CalledProcessError as e:
-                        logging.warning(f"Command failed: {e}")
+                    # run kill script on box
+                    subprocess.check_output(check_lock, shell=True, text=True)
                     subprocess.check_output(install_connect, shell=True, text=True)
                     subprocess.check_output(update_config, shell=True, text=True)
                     subprocess.check_output(restart_connect, shell=True, text=True)
@@ -87,6 +83,7 @@ def connect_ready(box_name, max_attempts, interval):
 
 api_key=get_api_key('admin')
 connect_version=get_connect_version()
+check_lock = "fuzzbucket-client ssh " + box_name + " " + ssh_options + " sudo ./check_lock.sh"
 install_connect = "fuzzbucket-client ssh " + box_name + " " + ssh_options + " sudo -E UNATTENDED=1 bash installer-ci.sh -d " + connect_version
 
 response = connect_ready(box_name, 20, 5)
